참조 수정

+++ src/main/java/com/example/zerobase/zerobaseminiassignment/service/FollowManageService.java
@@ -4,9 +4,12 @@ import com.example.zerobase.zerobaseminiassignment.common.MyJwtUtil;
 import com.example.zerobase.zerobaseminiassignment.model.FollowModel;
 import com.example.zerobase.zerobaseminiassignment.model.MemberModel;
 import com.example.zerobase.zerobaseminiassignment.repository.FollowRepository;
+import com.example.zerobase.zerobaseminiassignment.repository.MemberRepository;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+
+import java.util.Optional;
 
 @Slf4j
 @Service
@@ -14,14 +17,17 @@ public class FollowManageService {
 
     @Autowired
     private FollowRepository followRepository;
-
     @Autowired
-    private MemberManageService memberManageService;
+    private MemberRepository memberRepository;
+
     public boolean save(Long memberId) {
 
         try{
-            MemberModel followerMember = memberManageService.find(memberId);
-            followRepository.save(new FollowModel(MyJwtUtil.getMember(),followerMember));
+            Optional<MemberModel> followerMember = memberRepository.findById(memberId);
+            if(followerMember.isEmpty()){
+                return false;
+            }
+            followRepository.save(new FollowModel(MyJwtUtil.getMember(),followerMember.get()));
         }catch (Exception e){
             log.error(e.getMessage());
             return false;
@@ -32,9 +38,10 @@ public class FollowManageService {
     public boolean delete(Long memberId) {
 
         try{
-            MemberModel followerMember = memberManageService.find(memberId);
+            Optional<MemberModel> followerMember = memberRepository.findById(memberId);
+            if(followerMember.isEmpty()) return false;
             followRepository.deleteByFollowerMemberIdAndFollowingMemberId
-                    (MyJwtUtil.getMember(),followerMember);
+                    (MyJwtUtil.getMember(),followerMember.get());
         }catch (Exception e){
             log.error(e.getMessage());
             return false;
